ISSUE-114 Better handling of Generic/Specific Records
Allow producer with SpecificRecord and consumer with GenericRecord and vice versa.

This deprecates the use of the recordType in the protocol but we leave reading it, to keep compatbility.
Better handling of Generic/Specific Records

Convert SerDes class's to use Interface of SchemaRegistryClient instead of concreate class. 
Allow passing in of SchemaRegistryClient, useful for passing in a Mock instance of the client both for Registry testing and also end users testing.
Created a basic Mock Schema Registry Client, by default this wraps the ISchemaRegistry with default constuctor making an in memory version
Added Test Cases, for Specific to Generic and vice versa using a mocked reigstry.

NB. 
Should look to make the Mock Client usable by end users, before doing this need to raise concern that Client Interface != Server Interface really should align this, this because very apparent when trying to make a mocked client for testing, end users really will feel this pain.