ISSUE-486 Fix deadlocks in storage layer during concurrent calls to SR (#499)
In case of code path for adding schema metadata, we first do a get(schemaMetadataName), if the get returns a null then we add the new schemaMetadata else we return the schemaMetadataId from the get() call. The current transaction isolation level in SR is SERIALIZABLE, in this case, the get() call gets a shared lock and the add() call get an exclusive lock within a transaction. If two threads are invoking this API we might run into a situation where thread1 has obtained a shared lock and thread2 has also obtained the shared lock but thread1 can't get an exclusive lock because of a shared lock held by thread2 and thread2 can't get an exclusive lock because of a shared lock held by thread1. So, we run into a deadlock. The solution here would be to reduce the transaction isolation level to REPETABLE_READS for MySQL and POSTGRES and since ORACLE doesn't use this locking mechanism to in case of SERIALIZABLE, we can retain the current transaction isolation level for ORACLE.

Author: Sarvanan Raju <sraju@hortonworks.com>

Reviewers: @satishd,@priyank5485,@raju-saravanan

Closes #484 from raju-saravanan/deadlock