Backporting the fix which stablises the concurrent addition of schema versions (#518)
* ISSUE-517 Fix concurrent schema version adds.

When a schema version is to be added we query the database to get the latest schema version's version number, increment the value and it becomes the new schema version's version number. Currently, only one client is allowed to create a version and other concurrent requests are received an error and they should retry again. With this PR, other requests would wait till one of the requests finishes the update.
A new table "schema_locks" has been exposed as a part of this PR. For every, schema metadata that is added we add a new entry in "schema_locks" table and whenever a schema version is to be added a write lock is obtained on a row corresponding to that schema metadata on "schema_locks" table, then we can exclusively query the database to find the latest schema version. This would resolve the concurrent issue we run into while adding multiple schema versions parallel.

Changes were tested against Mysql, Postgres, Oracle 12c and Oracle 11gr2.

Closes #517

Author: Sarvanan Raju <sraju@hortonworks.com>

Reviewers: @satishd

Closes #516 from raju-saravanan/locks

*   MINOR : Plug in allocate_extent migration script for ORACLE