ISSUE-513: Fix storage manager to correctly handle the types in queries
Currently the JDBC storage manager user the value of the query parameter to guess the column type and uses this to construct the prepared statement. This is fragile and breaks when the value appears to be of a different type. For example a varchar column name and query string name="123".

We can leverage the JDBC result set to figure out the sql type and map it to the Schema.Type to fix this.

Author: Arun Mahadevan <arunm@apache.org>

Closes #514 from arunmahadevan/ISSUE-513