CDPD-12840 Relocate Jersey packages in schemaregistry-client in order to avoid conflict with other components.
Jersey is relying on jaxws-rs for sending REST requests. Unfortunately it is possible for other, external components to also use jaxws-rs, but a different version. If the two versions are on the same classpath then we see conflicts. In order to avoid conflicts, we have relocated the packages used by schemaregistry-client, so the package javax.ws.rs becomes com.hortonworks.registries.shaded.javax.ws.rs. This way, both jaxws-rs libraries are on the classpath, but due to the difference in package names they are not in conflict.

Set version to 0.9.1 in order to avoid conflicts with the snapshot repo.

Change-Id: I4053423c4c6163ce3d3bd8080c27bd529f98d88e