CDPD-18019 Remove unnecessary Spring dependencies from common-auth
schemaregistry-client is depending on common-auth and common-auth had dependencies on spring-core. As a consequence, the SR client also depends on spring-core which causes issues on the client side. After testing, we discovered this dependency on Spring is not needed at all, it's only used for testing. We only need to depend on spring-security for Kerberos authentication.
After removing spring from common-auth all tests continued to work and the unnecessary Spring dependencies have also been removed.

Change-Id: I0a6cb98573626c9875dbd11dbd20c2f4ae9d3161