CDPD-17969 Schema version cannot be forked when Ranger is enabled.
The issue is the server side expects to be provided with the name of the schema, while the UI does not provide it. We do not want to change the client side, so the solution is to use the schema ID instead of the name. The ID can be retrieved from the version ID, which has always been provided by the client.

Change-Id: I11d01949abeec5a1e11e862ab4c4447836a181ff