[CDPD-17737] Fix misused exception handling of SchemaRegistryResources
Jersey provides an exception mapper which enables us to handle exceptions to http response code translation in a central place (GenericExceptionMapper).
In SchemaRegistryResource and ConfluentSchemaRegistryCompatibleResource, try-catch blocks are deleted and resources are configured to use the GenericExceptionMapper.
In SchemaRegistryResource, createSchemaBranch method (endpoint /schemas/versionsById/{versionId}/branch) responds with 404 Not Found when SchemaNotFoundException is thrown (and not 400 Bad Request).
SchemaNotFoundException and SchemaBranchAlreadyExistsException changed to contain "entity".

Change-Id: I0de1961c529dba9dd3c634e867f3ca76880a6f46