[CDPD-31881] Schema Registry L1 test fails with socket timeout
Get atlas events with a SearchQuery modified to enable locking. When 'FOR UPDATE SKIP LOCKED' is supported by the DB, rows that get grabbed by a concurrent processor are skipped, or they wait for each other if no 'SKIP LOCKED' support.
Use READ_COMMITTED instead of SERIALIZABLE isolation level so that transactions are not aborted by the DB, and in a race the loser sees the winner's update of the 'processed' and 'failed' status so rows are not processed twice.

Change-Id: Ibd54bb34d1d56a50e291db3518ced02ab7ff6c41