CDPD-39741 - Move Schema Registry Ranger Plugin out from Schema Registry
This commit moves the ranger dependency out from schema-registry-authorizer. With this, the SR project will not have a dependency on Ranger. As some of the other modules depend on the Authorizer interface, these parts remain in the project. The relevant code parts were moved to a new project CDF/schema-registry-ranger-plugin.

Change-Id: Ic6e6a697b8888151efdef0aa2d185c4e1c3996df