CDPD-41798: Modify the code using Hadoop to use new methods.
In Apache Hadoop 3.3.3 some methods have been changed while they still exist in Hadoop 3.1.1. It is expected that at some point all components will upgrade to the newer version of Hadoop and then this code change will become inevitable. We are making the code change now so we don't need to fix it later.

Change-Id: Ibd0f32ca554e9ce12576ffd85594ec8c077b2de6