[CDPD-18529] Avro related code to be removed from SchemaRegistry
Removed code that duplicates org.apache.avro.SchemaCompatibility but still allow unions types having the "null" type first (like ["null", "long"]) to be treated as if they had a default null value implicitly. Avro version is constrained now to not have 3 different versions on the classpath. This in turn required going back to version 6.1.7 for the test-only confluent dependencies (upgraded due to a Netty related CVE) until we upgrade to Avro 1.11.

Change-Id: I50b67f313dca4e9f2167ed8f4a4f37f5d9117d3b