CDPD-49468: OAuth2Login thread safety
OAuth2Login was not thread safe. When accessed from multiple threads, it can renew the token in parallel, and overwrite the old token with a non-processed one.
Additionally, since there is no locking involved, and the authToken variable is not volatile, threads might access a stale token.
Change the authToken variable to be volatile, and the login() method only writes into it once. Parallel renewal is still possible, but it is a performance issue at worst, and not a functional one.

Change-Id: I2b051a80441b79aafd4b166beb19e53c8f0929ff