CDPD-54360: Support null data in Avro serde classes
The serializer now supports a new flag called null.passthrough.enabled (by default, false) - if enabled, the serializer handles null by returning null, without sending any schema to Schema Registry. The deserializer now also returns null if the data is null.
This can be used in conjunction with compact topics, where the value "null" has specific meaning (tombstone marker). Without the null passthrough, the serializer returns a byte array which contains the protocol and the version. This is in line with how avro serializes null (0 bytes), but since the returned data is not null, tombstone messages cannot be written with the serializer.

Change-Id: I5b9bad54726217b46aa69bcb943032d50ce45aa5