CDPD-49470, CDPD-49469: SchemaRegistryClient retries should handle each request as an attempt
Previously, SchemaRegistryClient rotated over each URL available through the UrlSelector in one try, then handed the control back to the RetryExecutor. This meant that
  - even if a no retry policy was installed, but there were multiple SR instances listed in the config, the client would try multiple times, without any backoff.
  - even if a retry policy was installed with a configured max attempt count, the actual number of attempts would at least max_attempts * number_of_sr_urls
Additionally, the logic of going through all available URLs assumed that each thread will get the URLs in round-robin, but in multi-threaded environments, this assumption is wrong. Other threads can mark specific URLs as failed, and depending on the UrlSelector implementation, assuming N configured SR URLs, the N+1th call to .select() might not return the URL returned in the 1st .select() call. This could potentially cause even more retries than intended, or an infinite cycle in most extreme cases.

All of the above was fixed by removing the "rotate over all URLs in one try" part of the logic, and now it is up to the RetryExecutor to perform the retries, respecting the max attempts and the backoff configurations.
At the same time, the default BackoffPolicy was changed to ExponentialBackoffPolicy to ensure that clients using multiple SR URLs with the default backoff policy will still get some sort of retries. For clients using a single SR URL with the default backoff policy, this will be a behavior change, as the default policy now will perform retries. In most cases, this should be an improvement in robustness, but if the new behavior is undesirable, clients must configure an explicit NOOPBackoffPolicy.

Change-Id: I417fc948cdefc50751a6d19addc34275ed709cd4