CDPD-58265: Workaround for jersey client default SSL socket factory concurrency issue
This is a workaround as described in https://github.com/eclipse-ee4j/jersey/issues/4332, which is related to a JDK bug: https://bugs.openjdk.org/browse/JDK-8160347
In short, HttpsURLConnection.getDefaultSSLSocketFactory is not thread safe, and it can return different objects on different calls. Because of this, concurrent access in Jersey clients can lead to an incorrect SSL socket configuration. An easy workaround is to make sure that HttpsURLConnection.getDefaultSSLSocketFactory is called once before any jersey access. Client applications can still encounter the issue if they start using a jersey client in their own code, but the SR client should be safe from the issue with this change.

Change-Id: I920d5b0a1bd6fd5851686ce1285ab2d98e6291bf