CDPD-60160: Convert SQL NUMERIC type to BigDecimal when querying data from DB
Previously it was converted to Boolean, Byte, Integer or Long based on precision.
This caused error in case of Oracle DB, where NUMBER is heavily used.
There were 2 issues with this approach:
1. The conversion occurred for some arbitrary precision values instead of ranges.
   E.g. NUMBER(10) -> Integer but NUMBER(8) -> Long
2. The largest number according to precision does not fit into its corresponding Java type.
   This results in silent truncation.
   E.g. NUMBER(1) to Boolean, NUMBER(10) to Integer
With this fix NUMERIC is converted to BigDecimal (which is possible according
to SQL-to-Java type mapping documentations) from which the value can be converted
to the proper type (Integer, Long) later. Exception is thrown if the conversion is not
possible without loss of precision.

Change-Id: I7488034b4eec853168823872834939003f42f0af
(cherry picked from commit 202edfbc66aeb028525560b985ec601dd60354e5)